requirements.txt
openshift-client
cryptography
